requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
pydantic == 2.11.4
joblib == 1.5.0
pandas == 2.2.3
Jinja2 == 3.1.6
python-multipart == 0.0.20
scikit-learn == 1.6.1
xgboost == 2.1.4
category-encoders == 2.6.4
SQLAlchemy == 2.0.30
psycopg2-binary == 2.9.9
python-dotenv == 1.1.0
redis== 4.5.3
fastapi-cache2 == 0.2.2